Fixing analyzer issues, targeting .NET 9 instead of .NET 6

+++ KGySoft.Drawing.ImagingTools/WinApi/Constants.cs
@@ -37,8 +37,8 @@ namespace KGySoft.Drawing.ImagingTools.WinApi
         internal const int WM_NCACTIVATE = 0x0086;
         internal const int WM_THEMECHANGED = 0x031A;
 
-        internal const int MA_ACTIVATEANDEAT = 2;
-        internal const int MA_ACTIVATE = 1;
+        internal const nint MA_ACTIVATEANDEAT = 2;
+        internal const nint MA_ACTIVATE = 1;
 
         internal const int SWP_NOSIZE = 0x0001;
         internal const int SWP_NOMOVE = 0x0002;
